+++ src/mlpack/tests/loss_functions_test.cpp
@@ -515,20 +515,20 @@ BOOST_AUTO_TEST_CASE(LogCoshLossTest)
 /**
  * Simple test for the Hinge Embedding loss function.
  */
-BOOST_AUTO_TEST_CASE(HingeEmbeddingLoss)
+BOOST_AUTO_TEST_CASE(HingeEmbeddingLossTest)
 {
   arma::mat input, target, output;
   double loss;
-  HingeEmbeddingLoss module;
+  HingeEmbeddingLoss<> module;
 
   // Test the Forward function. Loss should be 0 if input = target.
   input = arma::ones(10, 1);
   target = arma::ones(10, 1);
-  loss = module.Forward(std::move(input), std::move(target));
+  loss = module.Forward(input, target);
   BOOST_REQUIRE_EQUAL(loss, 0);
 
   // Test the Backward function for input = target.
-  module.Backward(std::move(input), std::move(target), std::move(output));
+  module.Backward(input, target, output);
   for (double el : output)
   {
     // For input = target we should get 0.0 everywhere.
@@ -541,11 +541,11 @@ BOOST_AUTO_TEST_CASE(HingeEmbeddingLoss)
   // Test the Forward function. Loss should be 0.71999997.
   input = arma::mat("0.1 0.8 0.6 0.0 0.5");
   target = arma::mat("0. 1.0 1.0 0. 0.");
-  loss = module.Forward(std::move(input), std::move(target));
+  loss = module.Forward(input, target);
   BOOST_REQUIRE_CLOSE(loss, 0.71999997, 1e-3);
 
   // Test the Backward function.
-  module.Backward(std::move(input), std::move(target), std::move(output));
+  module.Backward(input, target, output);
   BOOST_REQUIRE_CLOSE(arma::accu(output), -2, 1e-3);
   BOOST_REQUIRE_EQUAL(output.n_rows, input.n_rows);
   BOOST_REQUIRE_EQUAL(output.n_cols, input.n_cols);